backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "cryptography==42.0.5",
    "fastapi==0.110.0",
    "google-generativeai==0.8.3",
    "httpx==0.27.0",
    "motor==3.6.0",
    "openai==1.14.0",
    "pydantic==2.6.4",
    "pydantic-settings==2.2.1",
    "pyjwt==2.8.0",
    "python-dotenv==1.0.1",
    "uvicorn==0.29.0",
]
